+++ opt-env-conf/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+* Added a hidden `--render-reference-documentation` command.
+
 ## [0.3.0.0] - 2024-07-19
 
 ### Changed